Fix CommandWithoutConfig (#147) properly
Unfortunately, CommandWithoutConfig needs to access rootCmd in some way,
and because two packages can't import each other, this is the best I
could come up with
